Rename feed-search to feed-searches
