+++ client/state/data-layer/wpcom/read/feed/test/index.js
@@ -11,7 +11,7 @@ import freeze from 'deep-freeze';
 import {
 	requestFeedSearch,
 	receiveFeedSearch,
-} from 'state/reader/feed-search/actions';
+} from 'state/reader/feed-searches/actions';
 import {
 	initiateFeedSearch,
 	receiveFeeds,
