+++ client/state/reader/feed-searches/reducer.js
